Fix missing provider import in tax rate dialogs
