+++ lib/screens/simplified_quote_screen.dart
@@ -20,6 +20,8 @@ import '../widgets/added_products_list.dart';
 import '../widgets/custom_line_items_section.dart';
 import '../widgets/quote_levels_preview.dart';
 import '../controllers/quote_form_controller.dart';
+import '../dialogs/add_product_dialog.dart';
+import '../dialogs/custom_item_dialog.dart';
 
 class SimplifiedQuoteScreen extends StatefulWidget {
   final Customer customer;
@@ -283,7 +285,7 @@ class _SimplifiedQuoteScreenState extends State<SimplifiedQuoteScreen> {
   void _showAddCustomItemDialog() {
     showDialog(
       context: context,
-      builder: (context) => _CustomItemDialog(
+      builder: (context) => CustomItemDialog(
         onItemAdded: (item) {
           _controller.addCustomLineItem(item);
         },
@@ -385,7 +387,7 @@ class _SimplifiedQuoteScreenState extends State<SimplifiedQuoteScreen> {
   void _showAddProductDialog() {
     showDialog(
       context: context,
-      builder: (context) => _AddProductDialog(
+      builder: (context) => AddProductDialog(
         onProductAdded: (productItem) {
           _controller.addProduct(productItem);
         },
@@ -404,442 +406,3 @@ class _SimplifiedQuoteScreenState extends State<SimplifiedQuoteScreen> {
   }
 }
 
-// ADD PRODUCT DIALOG
-class _AddProductDialog extends StatefulWidget {
-  final Function(QuoteItem) onProductAdded;
-
-  const _AddProductDialog({required this.onProductAdded});
-
-  @override
-  State<_AddProductDialog> createState() => _AddProductDialogState();
-}
-
-class _AddProductDialogState extends State<_AddProductDialog> {
-  final _formKey = GlobalKey<FormState>();
-  final _quantityController = TextEditingController(text: '1.0');
-
-  Product? _selectedProduct;
-  ProductLevelPrice? _selectedLevel;
-  String _expandedCategory = '';
-
-  @override
-  void dispose() {
-    _quantityController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Dialog(
-      child: Container(
-        width: MediaQuery.of(context).size.width * 0.9,
-        height: MediaQuery.of(context).size.height * 0.7,
-        padding: const EdgeInsets.all(16),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            // Header
-            Row(
-              children: [
-                Icon(Icons.add_shopping_cart, color: Theme.of(context).primaryColor),
-                const SizedBox(width: 12),
-                const Expanded(
-                  child: Text(
-                    'Add Product to Quote',
-                    style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
-                  ),
-                ),
-                IconButton(
-                  icon: const Icon(Icons.close),
-                  onPressed: () => Navigator.pop(context),
-                ),
-              ],
-            ),
-            const Divider(),
-
-            // Product Selection
-            Expanded(
-              child: Form(
-                key: _formKey,
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    const Text(
-                      'Select Product:',
-                      style: TextStyle(fontWeight: FontWeight.bold),
-                    ),
-                    const SizedBox(height: 8),
-
-                    // Product Categories
-                    Expanded(
-                      child: Consumer<AppStateProvider>(
-                        builder: (context, appState, child) {
-                          final productsByCategory = _groupProductsByCategory(appState.products);
-
-                          return ListView(
-                            children: productsByCategory.entries.map((entry) {
-                              final category = entry.key;
-                              final products = entry.value;
-
-                              return Card(
-                                margin: const EdgeInsets.only(bottom: 8),
-                                child: Theme(
-                                  data: Theme.of(context).copyWith(dividerColor: Colors.transparent),
-                                  child: ExpansionTile(
-                                    title: Text(
-                                      '$category (${products.length})',
-                                      style: const TextStyle(fontWeight: FontWeight.w500),
-                                    ),
-                                    leading: Icon(_getCategoryIcon(category)),
-                                    initiallyExpanded: _expandedCategory == category,
-                                    onExpansionChanged: (expanded) {
-                                      setState(() {
-                                        _expandedCategory = expanded ? category : '';
-                                      });
-                                    },
-                                    children: products.map((product) {
-                                      final isSelected = _selectedProduct?.id == product.id;
-
-                                      return ListTile(
-                                        title: Text(product.name),
-                                        subtitle: Text(
-                                          '\$${product.unitPrice.toStringAsFixed(2)}/${product.unit}${product.enhancedLevelPrices.isNotEmpty
-                                                  ? ' (${product.enhancedLevelPrices.length} levels)'
-                                                  : ''}',
-                                        ),
-                                        leading: Radio<Product>(
-                                          value: product,
-                                          groupValue: _selectedProduct,
-                                          onChanged: (value) {
-                                            setState(() {
-                                              _selectedProduct = value;
-                                              _selectedLevel = null;
-                                            });
-                                          },
-                                        ),
-                                        selected: isSelected,
-                                        onTap: () {
-                                          setState(() {
-                                            _selectedProduct = product;
-                                            _selectedLevel = null;
-                                          });
-                                        },
-                                      );
-                                    }).toList(),
-                                  ),
-                                ),
-                              );
-                            }).toList(),
-                          );
-                        },
-                      ),
-                    ),
-
-                    const SizedBox(height: 16),
-
-                    // Level Selection
-                    if (_selectedProduct != null && _selectedProduct!.enhancedLevelPrices.isNotEmpty) ...[
-                      const Text(
-                        'Select Level:',
-                        style: TextStyle(fontWeight: FontWeight.bold),
-                      ),
-                      const SizedBox(height: 8),
-                      DropdownButtonFormField<ProductLevelPrice>(
-                        decoration: const InputDecoration(
-                          border: OutlineInputBorder(),
-                          prefixIcon: Icon(Icons.layers),
-                        ),
-                        value: _selectedLevel,
-                        hint: const Text('Choose level/option'),
-                        items: _selectedProduct!.enhancedLevelPrices.map((level) {
-                          return DropdownMenuItem(
-                            value: level,
-                            child: Text('${level.levelName} - \$${level.price.toStringAsFixed(2)}/${_selectedProduct!.unit}'),
-                          );
-                        }).toList(),
-                        onChanged: (level) {
-                          setState(() {
-                            _selectedLevel = level;
-                          });
-                        },
-                        validator: (value) => value == null ? 'Please select a level' : null,
-                      ),
-                      const SizedBox(height: 16),
-                    ],
-
-                    // Quantity Input
-                    TextFormField(
-                      controller: _quantityController,
-                      decoration: InputDecoration(
-                        labelText: 'Quantity',
-                        border: const OutlineInputBorder(),
-                        prefixIcon: const Icon(Icons.calculate),
-                        suffixText: _selectedProduct?.unit ?? '',
-                      ),
-                      keyboardType: const TextInputType.numberWithOptions(decimal: true),
-                      validator: (value) {
-                        if (value == null || value.isEmpty) return 'Please enter quantity';
-                        final qty = double.tryParse(value);
-                        if (qty == null || qty <= 0) return 'Enter a valid positive quantity';
-                        return null;
-                      },
-                    ),
-                  ],
-                ),
-              ),
-            ),
-
-            // Footer
-            const Divider(),
-            Row(
-              children: [
-                Expanded(
-                  child: OutlinedButton(
-                    onPressed: () => Navigator.pop(context),
-                    child: const Text('Cancel'),
-                  ),
-                ),
-                const SizedBox(width: 12),
-                Expanded(
-                  child: ElevatedButton(
-                    onPressed: _addProduct,
-                    child: const Text('Add to Quote'),
-                  ),
-                ),
-              ],
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Map<String, List<Product>> _groupProductsByCategory(List<Product> products) {
-    final grouped = <String, List<Product>>{};
-
-    for (final product in products) {
-      if (!product.isActive) continue;
-      if (product.pricingType == ProductPricingType.mainDifferentiator) continue;
-
-      final category = product.category;
-      grouped.putIfAbsent(category, () => []).add(product);
-    }
-
-    final sortedEntries = grouped.entries.toList()
-      ..sort((a, b) => a.key.compareTo(b.key));
-
-    final result = <String, List<Product>>{};
-    for (final entry in sortedEntries) {
-      entry.value.sort((a, b) => a.name.compareTo(b.name));
-      result[entry.key] = entry.value;
-    }
-
-    return result;
-  }
-
-  IconData _getCategoryIcon(String category) {
-    switch (category.toLowerCase()) {
-      case 'materials':
-      case 'roofing':
-        return Icons.roofing;
-      case 'gutters':
-        return Icons.water_drop;
-      case 'labor':
-        return Icons.engineering;
-      case 'flashing':
-        return Icons.flash_on;
-      default:
-        return Icons.category;
-    }
-  }
-
-  void _addProduct() {
-    if (!_formKey.currentState!.validate()) return;
-    if (_selectedProduct == null) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(content: Text('Please select a product')),
-      );
-      return;
-    }
-
-    // Determine price to use
-    double unitPrice;
-    String productName = _selectedProduct!.name;
-
-    if (_selectedProduct!.enhancedLevelPrices.isNotEmpty) {
-      if (_selectedLevel == null) {
-        ScaffoldMessenger.of(context).showSnackBar(
-          const SnackBar(content: Text('Please select a level for this product')),
-        );
-        return;
-      }
-      unitPrice = _selectedLevel!.price;
-      productName += ' (${_selectedLevel!.levelName})';
-    } else {
-      unitPrice = _selectedProduct!.unitPrice;
-    }
-
-    final quantity = double.parse(_quantityController.text);
-
-    final quoteItem = QuoteItem(
-      productId: _selectedProduct!.id,
-      productName: productName,
-      quantity: quantity,
-      unitPrice: unitPrice,
-      unit: _selectedProduct!.unit,
-      description: _selectedLevel?.description,
-    );
-
-    widget.onProductAdded(quoteItem);
-    Navigator.pop(context);
-
-    ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(
-        content: Text('Added $productName to all quote levels'),
-        backgroundColor: Colors.green,
-      ),
-    );
-  }
-
-}
-
-// NEW: Custom item dialog
-class _CustomItemDialog extends StatefulWidget {
-  final Function(CustomLineItem) onItemAdded;
-
-  const _CustomItemDialog({required this.onItemAdded});
-
-  @override
-  State<_CustomItemDialog> createState() => _CustomItemDialogState();
-}
-
-class _CustomItemDialogState extends State<_CustomItemDialog> {
-  final _formKey = GlobalKey<FormState>();
-  final _nameController = TextEditingController();
-  final _amountController = TextEditingController();
-  final _descriptionController = TextEditingController();
-  bool _isTaxable = true;
-
-  @override
-  Widget build(BuildContext context) {
-    return Dialog(
-      child: Container(
-        width: MediaQuery.of(context).size.width * 0.9,
-        padding: const EdgeInsets.all(20),
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            Row(
-              children: [
-                Icon(Icons.add_box, color: Colors.purple),
-                const SizedBox(width: 12),
-                const Expanded(
-                  child: Text(
-                    'Add Custom Line Item',
-                    style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
-                  ),
-                ),
-                IconButton(
-                  icon: const Icon(Icons.close),
-                  onPressed: () => Navigator.pop(context),
-                ),
-              ],
-            ),
-            const Divider(),
-            Form(
-              key: _formKey,
-              child: Column(
-                children: [
-                  TextFormField(
-                    controller: _nameController,
-                    decoration: const InputDecoration(
-                      labelText: 'Item Name',
-                      border: OutlineInputBorder(),
-                      prefixIcon: Icon(Icons.label),
-                    ),
-                    validator: (value) => value?.isEmpty == true ? 'Required' : null,
-                  ),
-                  const SizedBox(height: 16),
-                  TextFormField(
-                    controller: _amountController,
-                    decoration: const InputDecoration(
-                      labelText: 'Amount',
-                      border: OutlineInputBorder(),
-                      prefixIcon: Icon(Icons.attach_money),
-                      prefixText: '\$ ',
-                    ),
-                    keyboardType: const TextInputType.numberWithOptions(decimal: true),
-                    validator: (value) {
-                      if (value?.isEmpty == true) return 'Required';
-                      final amount = double.tryParse(value!);
-                      if (amount == null || amount < 0) return 'Enter valid amount';
-                      return null;
-                    },
-                  ),
-                  const SizedBox(height: 16),
-                  TextFormField(
-                    controller: _descriptionController,
-                    decoration: const InputDecoration(
-                      labelText: 'Description (Optional)',
-                      border: OutlineInputBorder(),
-                      prefixIcon: Icon(Icons.description),
-                    ),
-                    maxLines: 2,
-                  ),
-                  const SizedBox(height: 16),
-                  SwitchListTile(
-                    title: const Text('Taxable Item'),
-                    subtitle: const Text('Include this item in tax calculations'),
-                    value: _isTaxable,
-                    onChanged: (value) => setState(() => _isTaxable = value),
-                  ),
-                ],
-              ),
-            ),
-            const SizedBox(height: 20),
-            Row(
-              children: [
-                Expanded(
-                  child: OutlinedButton(
-                    onPressed: () => Navigator.pop(context),
-                    child: const Text('Cancel'),
-                  ),
-                ),
-                const SizedBox(width: 12),
-                Expanded(
-                  child: ElevatedButton(
-                    onPressed: _addCustomItem,
-                    style: ElevatedButton.styleFrom(backgroundColor: Colors.purple),
-                    child: const Text('Add Item'),
-                  ),
-                ),
-              ],
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  void _addCustomItem() {
-    if (!_formKey.currentState!.validate()) return;
-
-    final item = CustomLineItem(
-      name: _nameController.text,
-      amount: double.parse(_amountController.text),
-      description: _descriptionController.text.isEmpty ? null : _descriptionController.text,
-      isTaxable: _isTaxable,
-    );
-
-    widget.onItemAdded(item);
-    Navigator.pop(context);
-  }
-
-  @override
-  void dispose() {
-    _nameController.dispose();
-    _amountController.dispose();
-    _descriptionController.dispose();
-    super.dispose();
-  }
-}
